implementation/phase-05-agent-mcp/mcp-servers/mcp-kafka/requirements.txt
fastmcp==3.0.2
kafka-python-ng==2.2.3
uvicorn==0.37.0
